requirements.txt
numpy 
opencv-python 
tensorflow 
pillow 
scipy 
matplotlib 
scikit-learn
PyYAML
streamlit